requirements.txt
streamlit
google-generativeai
moviepy
openai
python-dotenv